tentative fix of preprocessing, probably breaks SER but seems to work with FITS

+++ src/core/arithm.h
@@ -22,7 +22,7 @@
 
 #include "core/siril.h"
 
-int soper(fits *a, double scalar, image_operator oper, gboolean conv_to_float);
+int soper(fits *a, float scalar, image_operator oper, gboolean conv_to_float);
 int imoper(fits *a, fits *b, image_operator oper, gboolean allow_32bits);
 int addmax(fits *a, fits *b);
 int siril_fdiv(fits *a, fits *b, float scalar, gboolean allow_32bits);
